Add secrets.json
This augments the `environment_secrets` repository rule to also emit
`secrets.json` alongside `secrets.bzl`. This allows application code to
parse the JSON file rather than the skylark file to retrieve all
secrets. Alternatively, the user could list the secrets in
`environment_secrets`, duplicate them in the `load()` statement, and
duplicate them in some sort of template expansion genrule to create a
file to be used at runtime.

NOTE: This also adds the latest version of skylib because it wasn't
available otherwise.